Adds e2e test for WelcomeScreen, adds onClick props, moves components to separate folders
