+++ package.json
@@ -15,6 +15,8 @@
     "@babel/preset-react": "7.6.3",
     "babel-jest": "24.9.0",
     "babel-loader": "8.0.6",
+    "enzyme": "3.10.0",
+    "enzyme-adapter-react-16": "1.15.1",
     "eslint": "5.13.0",
     "eslint-config-htmlacademy": "0.5.2",
     "eslint-plugin-react": "7.16.0",
