+++ src/components/app/app.jsx
@@ -1,11 +1,15 @@
 import React from "react";
 import PropTypes from "prop-types";
-import WelcomeScreen from './welcome-screen.jsx';
+import WelcomeScreen from '../welcome-screen/welcome-screen.jsx';
+
+const onStartGameClick = () => {
+  // TODO: implement
+};
 
 const App = (props) => {
   const {time, numErrors} = props;
 
-  return <WelcomeScreen time={time} numErrors={numErrors}
+  return <WelcomeScreen time={time} numErrors={numErrors} onClick={onStartGameClick}
   />;
 };
