+++ src/components/welcome-screen/welcome-screen.jsx
@@ -2,13 +2,13 @@ import React from "react";
 import PropTypes from "prop-types";
 
 const WelcomeScreen = (props) => {
-  const {time, numErrors} = props;
+  const {time, numErrors, onClick} = props;
 
   return <section className="welcome">
     <div className="welcome__logo">
       <img src="img/melody-logo.png" alt="Угадай мелодию" width="186" height="83"/>
     </div>
-    <button className="welcome__button"><span className="visually-hidden">Начать игру</span></button>
+    <button className="welcome__button" onClick={onClick}><span className="visually-hidden">Начать игру</span></button>
     <h2 className="welcome__rules-title">Правила игры</h2>
     <p className="welcome__text">Правила просты:</p>
     <ul className="welcome__rules-list">
@@ -22,6 +22,7 @@ const WelcomeScreen = (props) => {
 WelcomeScreen.propTypes = {
   time: PropTypes.number.isRequired,
   numErrors: PropTypes.number.isRequired,
+  onClick: PropTypes.func.isRequired
 };
 
 export default WelcomeScreen;
