+++ src/components/welcome-screen/welcome-screen.test.js
@@ -3,7 +3,7 @@ import renderer from "react-test-renderer";
 import WelcomeScreen from "./welcome-screen";
 
 it(`Welcome screen is rendered correctly after relaunch`, () => {
-  const tree = renderer.create(<WelcomeScreen time={0} numErrors={0}/>).toJSON();
+  const tree = renderer.create(<WelcomeScreen time={0} numErrors={0} onClick={jest.fn()}/>).toJSON();
 
   expect(tree).toMatchSnapshot();
 });
